ESLint: Move configs into configs dir
