+++ packages/eslint-config/configs/base.js
@@ -3,11 +3,11 @@ import tseslint from 'typescript-eslint';
 import importX from 'eslint-plugin-import-x';
 import prettier from 'eslint-config-prettier';
 
-import { checkFile } from './rules/check-file.js';
-import { preferArrow } from './rules/prefer-arrow.js';
-import { preferTemplate } from './rules/prefer-template.js';
-import { sortImports } from './rules/sort-imports.js';
-import compat from './utils/compat.js';
+import { checkFile } from '../rules/check-file.js';
+import { preferArrow } from '../rules/prefer-arrow.js';
+import { preferTemplate } from '../rules/prefer-template.js';
+import { sortImports } from '../rules/sort-imports.js';
+import compat from '../utils/compat.js';
 
 const baseConfig = tseslint.config(
   eslint.configs.recommended,
