+++ packages/eslint-config/configs/next.js
@@ -3,9 +3,9 @@ import tseslint from 'typescript-eslint';
 import globals from 'globals';
 
 import baseConfig from './base.js';
-import { nextRestrictedImports } from './rules/restricted-imports.js';
-import { nextCheckFile } from './rules/check-file.js';
-import compat from './utils/compat.js';
+import { nextRestrictedImports } from '../rules/restricted-imports.js';
+import { nextCheckFile } from '../rules/check-file.js';
+import compat from '../utils/compat.js';
 
 const nextConfig = tseslint.config(
   ...baseConfig,
